Move bonding curve module
