+++ curvesim/__init__.py
@@ -1,7 +1,7 @@
 """Package to simulate Curve pool."""
 __all__ = ["autosim", "bonding_curve", "order_book", "__version__", "__version_info__"]
 
-from ._bonding_curve import bonding_curve
 from ._order_book import order_book
 from .sim import autosim
+from .tools import bonding_curve
 from .version import __version__, __version_info__
