+++ curvesim/tools/__init__.py
@@ -1,0 +1,5 @@
+"Handy functions for the user interacting with Curvesim pools."
+
+__all__ = ["bonding_curve"]
+
+from .bonding_curve import bonding_curve
